please work

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,4 @@
-#include"200~function_pointers.h"
+#include "function_pointers.h"
 /**
  *array_iterator-function
  *@array: array of element to be printed
@@ -8,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-int i;
+size_t i;
 if (array && action && size > 0)
 {
 for (i = 0; i < size; i++)
